+++ apps/backoffice-v2/src/pages/Entity/components/Case/hooks/useDocuments/useDocumentsLogic.tsx
@@ -1,69 +1,23 @@
-import { t } from 'i18next';
-import { toast } from 'sonner';
 import { ComponentProps, useCallback, useRef, useState } from 'react';
 
 import { IDocumentsProps } from '../../interfaces';
 import { TransformWrapper } from 'react-zoom-pan-pinch';
 import { useCrop } from '@/common/hooks/useCrop/useCrop';
 import { DOWNLOAD_ONLY_MIME_TYPES } from '@/common/constants';
-import { useToggle } from '@/common/hooks/useToggle/useToggle';
 import { useFilterId } from '@/common/hooks/useFilterId/useFilterId';
 import { useTesseract } from '@/common/hooks/useTesseract/useTesseract';
 import { createArrayOfNumbers } from '@/common/utils/create-array-of-numbers/create-array-of-numbers';
 import { useStorageFileByIdQuery } from '@/domains/storage/hooks/queries/useStorageFileByIdQuery/useStorageFileByIdQuery';
-import { copyToClipboard } from '@/common/utils/copy-to-clipboard/copy-to-clipboard';
+import { useCustomerQuery } from '@/domains/customer/hook/queries/useCustomerQuery/useCustomerQuery';
 
-export const useDocuments = (documents: IDocumentsProps['documents']) => {
+export const useDocumentsLogic = (documents: IDocumentsProps['documents']) => {
   const initialImage = documents?.[0];
-  const {
-    crop,
-    isCropping,
-    onCrop,
-    cropImage,
-    toggleOnIsCropping,
-    toggleOffIsCropping,
-    onCancelCrop,
-  } = useCrop();
-  const [isLoadingOCR, , toggleOnIsLoadingOCR, toggleOffIsLoadingOCR] = useToggle(false);
+  const { data: customer } = useCustomerQuery();
+  const { crop, isCropping, onCrop, onCancelCrop } = useCrop();
   const selectedImageRef = useRef<HTMLImageElement>();
   const recognize = useTesseract();
   const filterId = useFilterId();
-  const onOcr = useCallback(async () => {
-    if (!isCropping) {
-      toggleOnIsCropping();
 
-      return;
-    }
-
-    toggleOnIsLoadingOCR();
-
-    try {
-      const croppedBase64 = await cropImage(selectedImageRef.current);
-      const result = await recognize(croppedBase64);
-      const text = result?.data?.text;
-
-      if (!text) {
-        throw new Error('No document OCR text found');
-      }
-
-      await copyToClipboard(text)();
-    } catch (err) {
-      console.error(err);
-
-      toast.error(t('toast:ocr_document_error'));
-    }
-
-    toggleOffIsLoadingOCR();
-    toggleOffIsCropping();
-  }, [
-    isCropping,
-    toggleOnIsLoadingOCR,
-    toggleOffIsLoadingOCR,
-    toggleOffIsCropping,
-    toggleOnIsCropping,
-    cropImage,
-    recognize,
-  ]);
   const skeletons = createArrayOfNumbers(4);
   const [selectedImage, setSelectedImage] = useState<{
     imageUrl: string;
@@ -114,11 +68,10 @@ export const useDocuments = (documents: IDocumentsProps['documents']) => {
     onCrop,
     onCancelCrop,
     isCropping,
-    onOcr,
+    isOCREnabled: !!customer?.features?.isDocumentOcrEnabled,
     selectedImageRef,
     initialImage,
     skeletons,
-    isLoadingOCR,
     selectedImage,
     onSelectImage,
     documentRotation,