+++ apps/kyb-app/src/common/providers/ThemeProvider/theme.context.ts
@@ -1,0 +1,4 @@
+import { IThemeContext } from '@/common/providers/ThemeProvider/types';
+import { createContext } from 'react';
+
+export const themeContext = createContext({} as IThemeContext);