+++ packages/workflow-core/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @ballerine/workflow-core
 
+## 0.6.51
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.39
+
+## 0.6.50
+
+### Patch Changes
+
+- bump
+- Updated dependencies
+  - @ballerine/common@0.9.38
+
+## 0.6.49
+
+### Patch Changes
+
+- version bump
+- Updated dependencies
+  - @ballerine/common@0.9.37
+
 ## 0.6.48
 
 ### Patch Changes